refactor: 로또 발행은 lottoManager의 역할로, lottoManager 클래스로 이동
