+++ __tests__/LottoManagerTest.js
@@ -1,57 +1,46 @@
-const MissionUtils = require('@woowacourse/mission-utils');
+const { mockQuestion, mockQuestions, EMPTY_CALLBACK } = require('../src/mock');
 const LottoManager = require('../src/LottoManager');
-
-const mockQuestions = (answers) => {
-  MissionUtils.Console.readLine = jest.fn();
-  answers.reduce(
-    (acc, input) =>
-      acc.mockImplementationOnce((question, callback) => {
-        callback(input);
-      }),
-    MissionUtils.Console.readLine,
-  );
-};
 
 describe('로또 매니저 클래스 테스트 / 당첨 번호 입력', () => {
   const lottoManager = new LottoManager();
 
   test('당첨 번호 입력이 모두 정수가 아니면 예외가 발생한다. ', () => {
-    const answer = ['1,2,3,4,5,a'];
+    const answer = '1,2,3,4,5,a';
 
-    mockQuestions(answer);
+    mockQuestion(answer);
 
     expect(() => {
-      lottoManager.inputWinningNumber(() => {});
+      lottoManager.inputWinningNumber(EMPTY_CALLBACK);
     }).toThrow('[ERROR]');
   });
 
   test('당첨 번호 입력이 1~45 사이의 입력이 아니면 예외가 발생한다.', () => {
-    const answer = ['1,2,3,4,5,80'];
+    const answer = '1,2,3,4,5,80';
 
-    mockQuestions(answer);
+    mockQuestion(answer);
 
     expect(() => {
-      lottoManager.inputWinningNumber(() => {});
+      lottoManager.inputWinningNumber(EMPTY_CALLBACK);
     }).toThrow('[ERROR]');
   });
 
   test('당첨 번호 입력의 개수가 6개가 아니면 예외가 발생한다. ', () => {
-    const answer = ['1,2,3,4,5,6,7'];
+    const answer = '1,2,3,4,5,6,7';
 
-    mockQuestions(answer);
+    mockQuestion(answer);
 
     expect(() => {
-      lottoManager.inputWinningNumber(() => {});
+      lottoManager.inputWinningNumber(EMPTY_CALLBACK);
     }).toThrow('[ERROR]');
   });
 
   test('당첨 번호 입력에 중복이 있으면 예외가 발생한다.', () => {
-    const answer = ['1,2,3,4,5,5'];
+    const answer = '1,2,3,4,5,5';
 
-    mockQuestions(answer);
+    mockQuestion(answer);
 
     expect(() => {
-      lottoManager.inputWinningNumber(() => {});
+      lottoManager.inputWinningNumber(EMPTY_CALLBACK);
     }).toThrow('[ERROR]');
   });
 });
@@ -60,12 +49,12 @@ describe('로또 게임 클래스 테스트 / 보너스 번호 입력', () => {
   const lottoManager = new LottoManager();
 
   test('보너스 번호가 1~45 사이의 입력이 아니라면 예외가 발생한다.', () => {
-    const answer = ['80'];
+    const answer = '80';
 
-    mockQuestions(answer);
+    mockQuestion(answer);
 
     expect(() => {
-      lottoManager.inputBouseNumber(() => {});
+      lottoManager.inputBouseNumber(EMPTY_CALLBACK);
     }).toThrow('[ERROR]');
   });
 
@@ -76,8 +65,18 @@ describe('로또 게임 클래스 테스트 / 보너스 번호 입력', () => {
     mockQuestions([winningNumber, answer]);
 
     expect(() => {
-      lottoManager.inputWinningNumber(() => {});
-      lottoManager.inputBouseNumber(() => {});
+      lottoManager.inputWinningNumber(EMPTY_CALLBACK);
+      lottoManager.inputBouseNumber(EMPTY_CALLBACK);
     }).toThrow('[ERROR]');
+  });
+});
+
+describe('로또 매니저 클래스 테스트 / 로또 발행', () => {
+  const lottoManager = new LottoManager();
+  test('원하는 로또 개수가 발행되는지 확인', () => {
+    const lottoCount = 2;
+    const lottos = lottoManager.publishLottos(lottoCount);
+
+    expect(lottos).toHaveLength(lottoCount);
   });
 });
